+++ rlgym_learn_logging/api/wrappers/logged_combined_reward.py
@@ -1,13 +1,14 @@
 """Module for the logged combined reward"""
 
-from typing import Generic, List, Dict, Any
+from typing import Any, Dict, Generic, List
 
 from rlgym.api import AgentID, StateType
-from void_logging.api.rewards.log import SEPARATOR
 
-from .log import Log
-from .logged_float import Logged
-from .logged_reward import LoggedReward
+from rlgym_learn_logging.api.rewards.log import SEPARATOR
+
+from ..rewards.log import Log
+from ..rewards.logged_float import Logged
+from ..rewards.logged_reward import LoggedReward
 
 
 class LoggedCombinedReward(