+++ rlgym_learn_logging/api/wrappers/reward_fn_wrapper.py
@@ -18,9 +18,9 @@ class RewardFnWrapper(LoggedWrapper[AgentID, StateType], Generic[AgentID, StateT
 
         if issubclass(type(reward_fn), LoggedReward):
             raise ValueError(
-                f"The wrapper {self.__class__.__name__} is meant to be used with \
-                the type {RewardFunction.__name__}, you used it with \
-                the type {type(reward_fn).__name__}"
+                f"The wrapper {self.__class__.__name__} is meant to be used with "
+                + f"the type {RewardFunction.__name__}, you used it with "
+                + f"the type {type(reward_fn).__name__}"
             )
         self._is_reward_fn_logged = True